+++ src/_utils.ts
@@ -1,6 +1,3 @@
-import { xchacha20poly1305 } from "@noble/ciphers/chacha";
-import { siv } from "@noble/ciphers/aes";
-
 type Awaited<T> = T extends Promise<infer U> ? Awaited<U> : T;
 type Promisified<T> = Promise<Awaited<T>>;
 
@@ -80,88 +77,6 @@ function base64Decode(input: string) {
 function base64Encode(input: Uint8Array) {
   if (globalThis.Buffer) {
     return Buffer.from(input).toString("base64");
-  }
-  return globalThis.btoa(String.fromCodePoint(...input));
-}
-
-// Encryption
-
-// Use only for GCM-SIV, due to nonce-misuse-resistance. We need deterministic keys.
-const predefinedSivNonce = "ThtnxLK9eCF4OLMg";
-const encryptionPrefix = "$enc:";
-
-export interface StorageValueEnvelope {
-  nonce: string;
-  encryptedValue: string;
-}
-
-export function encryptStorageValue(
-  storageValue: any,
-  key: string,
-  raw?: boolean,
-): StorageValueEnvelope {
-  const cryptoKey = genBytesFromBase64(key);
-  const nonce = globalThis.crypto.getRandomValues(new Uint8Array(24));
-  const chacha = xchacha20poly1305(cryptoKey, nonce);
-  const encryptedValue = chacha.encrypt(
-    raw ? storageValue : new TextEncoder().encode(storageValue),
-  );
-  return {
-    encryptedValue: genBase64FromBytes(new Uint8Array(encryptedValue)),
-    nonce: genBase64FromBytes(nonce),
-  };
-}
-
-export function decryptStorageValue<T>(
-  storageValue: StorageValueEnvelope,
-  key: string,
-  raw?: boolean,
-): T {
-  const { encryptedValue, nonce } = storageValue;
-  const cryptoKey = genBytesFromBase64(key);
-  const chacha = xchacha20poly1305(cryptoKey, genBytesFromBase64(nonce));
-  const decryptedValue = chacha.decrypt(genBytesFromBase64(encryptedValue));
-  return raw ? (decryptedValue as T) : (new TextDecoder().decode(decryptedValue) as T);
-}
-
-export function encryptStorageKey(storageKey: string, key: string): string {
-  const cryptoKey = genBytesFromBase64(key);
-  const gcmSiv = siv(cryptoKey, genBytesFromBase64(predefinedSivNonce));
-  const encryptedKey = gcmSiv.encrypt(new Uint8Array(new TextEncoder().encode(storageKey)));
-  return encryptionPrefix + genBase64FromBytes(encryptedKey, true);
-}
-
-export function decryptStorageKey(encryptedKey: string, key: string): string {
-  const cryptoKey = genBytesFromBase64(key);
-  const gcmSiv = siv(cryptoKey, genBytesFromBase64(predefinedSivNonce));
-  const decryptedKey = gcmSiv.decrypt(
-    genBytesFromBase64(encryptedKey.slice(encryptionPrefix.length), true),
-  );
-  return new TextDecoder().decode(decryptedKey);
-}
-
-// Base64 utilities - Waiting for https://github.com/unjs/knitwork/pull/83 // TODO: Replace with knitwork imports as soon as PR is merged
-
-function genBytesFromBase64(input: string, urlSafe?: boolean) {
-  if (urlSafe) {
-    input = input.replace(/-/g, "+").replace(/_/g, "/");
-    const paddingLength = input.length % 4;
-    if (paddingLength === 2) {
-      input += "==";
-    } else if (paddingLength === 3) {
-      input += "=";
-    }
-  }
-  return Uint8Array.from(globalThis.atob(input), (c) => c.codePointAt(0) as number);
-}
-
-function genBase64FromBytes(input: Uint8Array, urlSafe?: boolean) {
-  if (urlSafe) {
-    return globalThis
-      .btoa(String.fromCodePoint(...input))
-      .replace(/\+/g, "-")
-      .replace(/\//g, "_")
-      .replace(/=+$/, "");
   }
   return globalThis.btoa(String.fromCodePoint(...input));
 }
