+++ src/utils.ts
@@ -1,12 +1,3 @@
-import destr from "destr";
-import {
-  decryptStorageKey,
-  decryptStorageValue,
-  encryptStorageKey,
-  encryptStorageValue,
-  stringify,
-  type StorageValueEnvelope,
-} from "./_utils.ts";
 import type { Storage, StorageValue, TransactionOptions } from "./types.ts";
 
 type StorageKeys = Array<keyof Storage>;
@@ -87,95 +78,6 @@ export function prefixStorage<T extends StorageValue>(
   return nsStorage;
 }
 
-export function encryptedStorage<T extends StorageValue>(
-  storage: Storage<T>,
-  encryptionKey: string,
-  encryptKeys = false,
-): Storage<T> {
-  const encStorage: Storage<T> = { ...storage };
-
-  const getStoredKey = (key: string) =>
-    encryptKeys ? encryptStorageKey(normalizeKey(key), encryptionKey) : key;
-
-  const getItemOptions = (
-    item: string | { key: string; options?: TransactionOptions },
-    commonOptions?: TransactionOptions,
-  ) => {
-    return typeof item === "string" || !item.options
-      ? commonOptions
-      : { ...commonOptions, ...item.options };
-  };
-
-  encStorage.hasItem = (key, ...args) => storage.hasItem(getStoredKey(key), ...args);
-
-  encStorage.getItem = (async (key, ...args) => {
-    const value = await storage.getItem<StorageValueEnvelope>(getStoredKey(key), ...args);
-    return value === null ? null : destr(decryptStorageValue<string>(value, encryptionKey));
-  }) as Storage<T>["getItem"];
-
-  encStorage.getItems = (async (items, commonOptions) => {
-    return Promise.all(
-      items.map(async (item) => {
-        const key = typeof item === "string" ? item : item.key;
-        return {
-          key: normalizeKey(key),
-          value: await encStorage.getItem(key, getItemOptions(item, commonOptions)),
-        };
-      }),
-    );
-  }) as Storage<T>["getItems"];
-
-  encStorage.getItemRaw = async (key, ...args) => {
-    const value = await storage.getItem<StorageValueEnvelope>(getStoredKey(key), ...args);
-    return value === null ? null : decryptStorageValue(value, encryptionKey, true);
-  };
-
-  encStorage.setItem = async (key, value, ...args) => {
-    const encryptedValue = encryptStorageValue(stringify(value), encryptionKey);
-    return storage.setItem(getStoredKey(key), encryptedValue as unknown as T, ...args);
-  };
-
-  encStorage.setItems = async (items, commonOptions) => {
-    await Promise.all(
-      items.map((item) =>
-        encStorage.setItem(item.key, item.value, getItemOptions(item, commonOptions)),
-      ),
-    );
-  };
-
-  encStorage.setItemRaw = async (key, value, ...args) => {
-    const encryptedValue = encryptStorageValue(value, encryptionKey, true);
-    return storage.setItem(getStoredKey(key), encryptedValue as unknown as T, ...args);
-  };
-
-  encStorage.removeItem = (key, ...args) => storage.removeItem(getStoredKey(key), ...args);
-  encStorage.setMeta = (key, ...args) => storage.setMeta(getStoredKey(key), ...args);
-  encStorage.getMeta = (key, ...args) => storage.getMeta(getStoredKey(key), ...args);
-  encStorage.removeMeta = (key, ...args) => storage.removeMeta(getStoredKey(key), ...args);
-
-  encStorage.getKeys = async (base, ...args) => {
-    if (!encryptKeys) {
-      return storage.getKeys(base, ...args);
-    }
-
-    const normalizedBase = normalizeKey(base);
-    const keys = await storage.getKeys("", ...args);
-    const decryptedKeys = keys.map((key) => decryptStorageKey(key, encryptionKey));
-
-    return normalizedBase
-      ? decryptedKeys.filter((key) => key.startsWith(normalizedBase))
-      : decryptedKeys;
-  };
-
-  encStorage.keys = encStorage.getKeys;
-  encStorage.get = (key, opts = {}) => encStorage.getItem(key, opts);
-  encStorage.set = (key, value, opts = {}) => encStorage.setItem(key, value, opts);
-  encStorage.has = (key, opts = {}) => encStorage.hasItem(key, opts);
-  encStorage.del = (key, opts = {}) => encStorage.removeItem(key, opts);
-  encStorage.remove = (key, opts = {}) => encStorage.removeItem(key, opts);
-
-  return encStorage;
-}
 
 export function normalizeKey(key?: string): string {
   if (!key) {
